distribute code across bivar_mixer and gsa_mixer
