+++ precimed/gsa_mixer/utils_test.py
@@ -1,23 +1,9 @@
 import numpy as np
 import logging
-import collections
-from .utils_obsolete import UnivariateParams_obsolete
-from scipy.sparse import csr_matrix, coo_matrix, csc_matrix
-import pytest
+from scipy.sparse import csr_matrix
 
-from .utils import UnivariateParametrization_natural_axis
-from .utils import UnivariateParametrization_constPI
-from .utils import UnivariateParametrization_constPI_constSIG2BETA
 from .utils import AnnotUnivariateParams
 from .utils import AnnotUnivariateParametrization
-from .utils import BivariateParams
-from .utils import _calculate_univariate_uncertainty
-from .cli import apply_univariate_fit_sequence
-from .cli import apply_bivariate_fit_sequence
-
-from .utils import _log_exp_converter
-from .utils import _logit_logistic_converter
-from .utils import _arctanh_tanh_converter
 
 def get_null_params(libbgmg):
     return AnnotUnivariateParams(libbgmg=libbgmg, s=0, l=0, sig2_zeroL=0)
@@ -185,117 +171,3 @@ def test_AnnotUnivariateParamsWithGenemat():
     assert(np.all(np.isclose(gene_h2, np.array(expected_h2).astype(np.float32))))
     assert(np.all(np.isclose(gene_snps, np.array(expected_snps).astype(np.float32))))
     assert(np.all(np.isclose(gene_enrich, np.array(expected_enrich).astype(np.float32))))
-
-def check_optimize_result(optimize_result, expected_cost_df):
-    assert(optimize_result['message'] in ['Optimization terminated successfully.', 'Not a bracketing interval.', 'Bracketing values (xa, xb, xc) do not fulfill this requirement: (f(xb) < f(xa)) and (f(xb) < f(xc))'])
-    assert(optimize_result['nfev'] >= 4)
-    assert(optimize_result['nit'] >= 1)
-    assert(np.isfinite(optimize_result['AIC']))
-    assert(np.isfinite(optimize_result['BIC']))
-    if expected_cost_df is not None:
-        assert(optimize_result['cost_df'] == expected_cost_df)
-
-# py.test precimed/mixer/utils_test.py  -k test_univariate_fit_diffevo
-def test_univariate_fit_diffevo():
-    libbgmg_mock = LibBgmgMock(num_snp=10, num_tag=5, make_weights=True)
-    args = collections.namedtuple('Args', ['diffevo_fast_repeats', 'seed', 'load_params_file'])._make((3, 123, ''))
-    params, optimize_result_sequence = apply_univariate_fit_sequence(args, libbgmg_mock, ['diffevo-fast'], init_params=None, trait=1)
-    print(params, optimize_result_sequence)
-    assert(isinstance(params, AnnotUnivariateParams))
-    check_optimize_result(optimize_result_sequence[0][1], expected_cost_df=3)
-
-# py.test precimed/mixer/utils_test.py  -k test_univariate_fit_neldermead
-def test_univariate_fit_neldermead():
-    libbgmg_mock = LibBgmgMock(num_snp=10, num_tag=5, make_weights=True)
-    args = collections.namedtuple('Args', ['diffevo_fast_repeats', 'seed', 'load_params_file'])._make((3, 123, ''))
-    params, optimize_result_sequence = apply_univariate_fit_sequence(args, libbgmg_mock, ['neldermead-fast'], init_params=get_params(libbgmg_mock), trait=1)
-    print(params, optimize_result_sequence)
-    assert(isinstance(params, AnnotUnivariateParams))
-    check_optimize_result(optimize_result_sequence[0][1], expected_cost_df=3)
-
-# py.test precimed/mixer/utils_test.py  -k test_univariate_fit_inflation
-def test_univariate_fit_inflation():
-    libbgmg_mock = LibBgmgMock(num_snp=10, num_tag=5, make_weights=True)
-    args = collections.namedtuple('Args', ['diffevo_fast_repeats', 'seed', 'load_params_file'])._make((3, 123, ''))
-    params, optimize_result_sequence = apply_univariate_fit_sequence(args, libbgmg_mock, ['inflation'], init_params=get_params(libbgmg_mock), trait=1)
-    print(params, optimize_result_sequence)
-    assert(isinstance(params, AnnotUnivariateParams))
-    check_optimize_result(optimize_result_sequence[0][1], expected_cost_df=1)
-
-# py.test precimed/mixer/utils_test.py  -k test_univariate_fit_infinitesimal
-def test_univariate_fit_infinitesimal():
-    libbgmg_mock = LibBgmgMock(num_snp=10, num_tag=5, make_weights=True)
-    args = collections.namedtuple('Args', ['diffevo_fast_repeats', 'seed', 'load_params_file'])._make((3, 123, ''))
-    params, optimize_result_sequence = apply_univariate_fit_sequence(args, libbgmg_mock, ['infinitesimal'], init_params=get_params(libbgmg_mock), trait=1)
-    print(params, optimize_result_sequence)
-    assert(isinstance(params, AnnotUnivariateParams))
-    check_optimize_result(optimize_result_sequence[0][1], expected_cost_df=2)
-
-# py.test precimed/mixer/utils_test.py  -k test_univariate_uncertainty
-def test_univariate_uncertainty():
-    libbgmg_mock = LibBgmgMock(num_snp=10, num_tag=5, make_weights=True)    
-    num_ci_samples = 5
-    params = get_true_params(libbgmg_mock)
-    NCKoef = params.find_nckoef()
-    ci, ci_sample = _calculate_univariate_uncertainty(params, UnivariateParametrization_natural_axis(params, libbgmg_mock, trait=1), 0.95, num_samples=num_ci_samples, NCKoef=NCKoef)
-    #print(ci, ci_sample)
-    for k in ci.keys():
-        for stat in ci[k].keys():
-            assert(np.isfinite(ci[k][stat]))
-    assert(len(ci_sample) == num_ci_samples)
-
-# py.test precimed/mixer/utils_test.py  -k test_bivariate_fit_sequence
-def test_bivariate_fit_sequence():
-    libbgmg_mock = LibBgmgMock(num_snp=10, num_tag=5, make_weights=True)
-    args = collections.namedtuple('Args', ['diffevo_fast_repeats', 'seed', 'analysis', 'load_params_file', 'trait1_params_file', 'trait2_params_file'])._make((3, 123, 'fit2', None, get_params(libbgmg_mock), get_params(libbgmg_mock)))
-    params, _, _, optimize_result_sequence = apply_bivariate_fit_sequence(args, libbgmg_mock, ['diffevo-fast', 'neldermead-fast', 'brute1', 'brent1'])
-    #print(params, optimize_result_sequence)
-    assert(isinstance(params, BivariateParams))
-    for optimize_result in optimize_result_sequence:
-        check_optimize_result(optimize_result[1], expected_cost_df=None)
-    assert(len(optimize_result_sequence) == 4)
-
-    args = collections.namedtuple('Args', ['analysis', 'load_params_file'])._make(('test2', params))
-    params, _, _, optimize_result_sequence = apply_bivariate_fit_sequence(args, libbgmg_mock, ['load', 'inflation'])
-    print(params, optimize_result_sequence)
-    for optimize_result in optimize_result_sequence:
-        check_optimize_result(optimize_result[1], expected_cost_df=None)
-    assert(len(optimize_result_sequence) == 3)  # inflation is fitted for sig2_zeroA(trait1), sig2_zeroA(trait2), and rho_zero
-
-# py.test precimed/mixer/utils_test.py  -k test_converters
-def test_converters():
-    epsval = np.finfo(float).eps
-    minval = np.finfo(float).min
-    maxval = np.finfo(float).max
-
-    def isclose(a, b, rel_tol=1e-09, abs_tol=2*epsval):
-        return abs(a-b) <= max(rel_tol * max(abs(a), abs(b)), abs_tol)
-
-    tests = [(_log_exp_converter,        0.0, maxval, [0.00001, 0.001, 1.2, 1000, 10000]), 
-             (_logit_logistic_converter, 0.0, 1.0,   [0.00001, 0.001, 0.1, 0.9, 0.999, 0.99999]),
-             (_arctanh_tanh_converter,    -1.0, 1.0,  [-0.9999, -0.991, -0.9, 0.9, 0.999, 0.99999])]
-
-    for func, limLow, limHigh, values in tests:
-        for index, val in enumerate([limLow] + values + [limHigh]):
-            #print('{}: {} -> {} -> {}'.format(func, val, func(val, False), func(func(val, False), True)))
-            assert np.isfinite(func(val, False)), 'Error in {}({})'.format(func, val)
-            assert isclose(val, func(func(val, False), True)), 'Error in {}({})'.format(func, val)
-            isLimit = (index==0) or (index==(len(values)+ 1))
-            assert abs(func(val, False)) < (1000 if isLimit else 20), '{}({}) is too large ({})'.format(func, val, func(val, False))
-            assert abs(func(val, False)) > (0.001 if isLimit else 0.1), '{}({}) is too small ({})'.format(func, val, func(val, False))
-        assert isclose(func(-10000, True), limLow), '{} vs {}'.format(func(-10000, True), limLow)
-        assert isclose(func( 10000, True), limHigh), '{} vs {}'.format(func(10000, True), limHigh)
-
-# py.test precimed/mixer/utils_test.py  -k test_parametrizations
-@pytest.mark.parametrize("UnivariateParametrization,expected_vec",
-    [(UnivariateParametrization_natural_axis, [0.20701416938432612, -9.210340371976182, -6.906754778648554]),
-     (UnivariateParametrization_constPI, [-9.210340371976182, 0.20701416938432612]), 
-     (UnivariateParametrization_constPI_constSIG2BETA, [0.20701416938432612])])
-def test_parametrizations(UnivariateParametrization, expected_vec):
-    params = UnivariateParams_obsolete(0.001, 1e-4, 1.23)
-    pp=UnivariateParametrization(params, None, 1)
-    vec=pp.params_to_vec(params)
-    assert(np.all(np.isclose(vec, expected_vec)))
-    params2 = pp.vec_to_params(vec)
-    assert(np.all(np.isclose([params.pi, params.sig2_beta, params.sig2_zeroA],
-                             [params2.pi, params2.sig2_beta, params2.sig2_zeroA])))
